Reorganize tests & imports
